+++ django_auth_adfs/config.py
@@ -13,6 +13,7 @@ from django.contrib.auth import REDIRECT_FIELD_NAME
 from django.contrib.auth import get_user_model
 from django.core.exceptions import ImproperlyConfigured
 from django.http import QueryDict
+from django.utils.module_loading import import_string
 
 try:
     from django.urls import reverse
@@ -23,13 +24,34 @@ logger = logging.getLogger("django_auth_adfs")
 
 AZURE_AD_SERVER = "login.microsoftonline.com"
 
+DEFAULT_SETTINGS_CLASS = 'django_auth_adfs.config.Settings'
+
 
 class ConfigLoadError(Exception):
     pass
 
 
+def _get_auth_adfs_settings():
+    """
+    Get the AUTH_ADFS setting from the Django settings.
+    """
+    if not hasattr(django_settings, "AUTH_ADFS"):
+        msg = "The configuration directive 'AUTH_ADFS' was not found in your Django settings"
+        raise ImproperlyConfigured(msg)
+    return django_settings.AUTH_ADFS
+
+
 class Settings(object):
+    """
+    Settings implementation reading from the Django settings.
+    """
+
     def __init__(self):
+        # which settings class to actually use - might be a replacement of this class
+        # unused in this class, but added here to contain all possible settings
+        # in the AUTH_ADFS django setting
+        self.SETTINGS_CLASS = DEFAULT_SETTINGS_CLASS
+
         # Set defaults
         self.AUDIENCE = None  # Required
         self.BOOLEAN_CLAIM_MAPPING = {}
@@ -66,36 +88,34 @@ class Settings(object):
             "TOKEN_PATH": "This setting is automatically loaded from ADFS.",
         }
 
-        if not hasattr(django_settings, "AUTH_ADFS"):
-            msg = "The configuration directive 'AUTH_ADFS' was not found in your Django settings"
-            raise ImproperlyConfigured(msg)
+        _auth_adfs_settings = _get_auth_adfs_settings()
 
         # Handle deprecated settings
         for setting, message in deprecated_settings.items():
-            if setting in django_settings.AUTH_ADFS:
+            if setting in _auth_adfs_settings:
                 warnings.warn("Setting {} is deprecated and it's value was ignored. {}".format(setting, message),
                               DeprecationWarning)
-                del django_settings.AUTH_ADFS[setting]
+                del _auth_adfs_settings[setting]
 
-        if "CERT_MAX_AGE" in django_settings.AUTH_ADFS:
-            django_settings.AUTH_ADFS["CONFIG_RELOAD_INTERVAL"] = django_settings.AUTH_ADFS["CERT_MAX_AGE"]
+        if "CERT_MAX_AGE" in _auth_adfs_settings:
+            _auth_adfs_settings["CONFIG_RELOAD_INTERVAL"] = _auth_adfs_settings["CERT_MAX_AGE"]
             warnings.warn('Setting CERT_MAX_AGE has been renamed to CONFIG_RELOAD_INTERVAL. The value was copied.',
                           DeprecationWarning)
-            del django_settings.AUTH_ADFS["CERT_MAX_AGE"]
+            del _auth_adfs_settings["CERT_MAX_AGE"]
 
-        if "GROUP_CLAIM" in django_settings.AUTH_ADFS:
-            django_settings.AUTH_ADFS["GROUPS_CLAIM"] = django_settings.AUTH_ADFS["GROUP_CLAIM"]
+        if "GROUP_CLAIM" in _auth_adfs_settings:
+            _auth_adfs_settings["GROUPS_CLAIM"] = _auth_adfs_settings["GROUP_CLAIM"]
             warnings.warn('Setting GROUP_CLAIM has been renamed to GROUPS_CLAIM. The value was copied.',
                           DeprecationWarning)
-            del django_settings.AUTH_ADFS["GROUP_CLAIM"]
+            del _auth_adfs_settings["GROUP_CLAIM"]
 
-        if "RESOURCE" in django_settings.AUTH_ADFS:
-            django_settings.AUTH_ADFS["RELYING_PARTY_ID"] = django_settings.AUTH_ADFS["RESOURCE"]
-            del django_settings.AUTH_ADFS["RESOURCE"]
+        if "RESOURCE" in _auth_adfs_settings:
+            _auth_adfs_settings["RELYING_PARTY_ID"] = _auth_adfs_settings["RESOURCE"]
+            del _auth_adfs_settings["RESOURCE"]
 
-        if "TENANT_ID" in django_settings.AUTH_ADFS:
+        if "TENANT_ID" in _auth_adfs_settings:
             # If a tenant ID was set, switch to Azure AD mode
-            if "SERVER" in django_settings.AUTH_ADFS:
+            if "SERVER" in _auth_adfs_settings:
                 raise ImproperlyConfigured("The SERVER cannot be set when TENANT_ID is set.")
             self.SERVER = AZURE_AD_SERVER
             self.USERNAME_CLAIM = "upn"
@@ -105,7 +125,7 @@ class Settings(object):
                                   "email": "email"}
 
         # Overwrite defaults with user settings
-        for setting, value in django_settings.AUTH_ADFS.items():
+        for setting, value in _auth_adfs_settings.items():
             if hasattr(self, setting):
                 setattr(self, setting, value)
             else:
@@ -302,5 +322,9 @@ class ProviderConfig(object):
         return self.end_session_endpoint
 
 
-settings = Settings()
+# Allow custom Setting providers via Django settings.
+AUTH_ADFS = _get_auth_adfs_settings()
+SettingsClass = import_string(AUTH_ADFS.get('SETTINGS_CLASS', DEFAULT_SETTINGS_CLASS))
+
+settings = SettingsClass()
 provider_config = ProviderConfig()