+++ tests/custom_config.py
@@ -1,0 +1,6 @@
+class Settings(object):
+    RETRIES = 1
+    CA_BUNDLE = False
+
+    def __init__(self):
+        self.SERVER = 'custom-server'